rl-dropdown-container: Inline `rl-dropdown-component` mixin
